+++ src/components/Settings.vue
@@ -1,25 +1,3 @@
-<!--
- - @copyright Copyright (c) 2019 John Molakvoæ <skjnldsv@protonmail.com>
- -
- - @author John Molakvoæ <skjnldsv@protonmail.com>
- -
- - @license AGPL-3.0-or-later
- -
- - This program is free software: you can redistribute it and/or modify
- - it under the terms of the GNU Affero General Public License as
- - published by the Free Software Foundation, either version 3 of the
- - License, or (at your option) any later version.
- -
- - This program is distributed in the hope that it will be useful,
- - but WITHOUT ANY WARRANTY; without even the implied warranty of
- - MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- - GNU Affero General Public License for more details.
- -
- - You should have received a copy of the GNU Affero General Public License
- - along with this program. If not, see <http://www.gnu.org/licenses/>.
- -
- -->
-
 <template>
   <div>
     <NcAppSettingsDialog
@@ -70,6 +48,14 @@
           type="switch"
         >
           {{ t('memories', 'Show full file path in sidebar') }}
+        </NcCheckboxRadioSwitch>
+
+        <NcCheckboxRadioSwitch
+          :checked.sync="config.metadata_in_slideshow"
+          @update:checked="updateMetadataInSlideshow"
+          type="switch"
+        >
+          {{ t('memories', 'Show metadata in slideshow') }}
         </NcCheckboxRadioSwitch>
 
         <div class="radio-group">
@@ -328,6 +314,10 @@ export default defineComponent({
 
     async updateSidebarFilepath() {
       await this.updateSetting('sidebar_filepath', 'sidebarFilepath');
+    },
+
+    async updateMetadataInSlideshow() {
+      await this.updateSetting('metadata_in_slideshow', 'metadataInSlideshow');
     },
 
     // Folders settings